% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.property.R
\name{assign.properties}
\alias{assign.properties}
\title{Function to assign properties to an expression matrix}
\usage{
assign.properties(exprs, genes, genelists = c("S4", "TNBCType"), n = 1000,
  seed = 123456, mc.cores = 1)
}
\arguments{
\item{exprs}{An expression matrix. Rows correspont to genes, columns to samples. Genes with multiple probes should be collapsed to a single row e.g. using most variable probe.}

\item{genes}{A vector with gene symbols for exprs. Should be the same length as the number of rows in exprs.}

\item{genelists}{A vector with either S4, TNBCType, or both to specify which genelists to use.}

\item{n}{An integer value specifying the number of random samples to generate.}

\item{seed}{An integer value specifying a seed for the random ranks function. Defauly value is 123456.}

\item{mc.cores}{An integer specifying how many cores to use. Default value is 1.}
}
\description{
This is a function to assign stromal property and TNBCType generative property levels to a TNBC dataset
}

