% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds.GSE13159}
\alias{ds.GSE13159}
\title{Microarray data of Leukemia after preprocessing(GSE13159)}
\format{
A list of two components: Expression matrix with 105 samples(Row) and 2000 genes(columns); a vector of class label for samples.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE13159}
}
\usage{
ds.GSE13159
}
\description{
Expression matrix with 105 samples(Row) and 2000 genes(columns). Originally there are 54,676 probesets in each dataset
(GSE6891, GSE17855 and GSE13159) and we remove the probesets with missing values and select the probesets with the largest interquartile
range to represent the gene if multiple probesets are mapped to the same gene. 20,192 unique genes
remained for every study after this preprocessing. Furthermore, for each study, we transform data
to log scale and only keep the top 10,000 genes with the largest mean expression level (i.e. filter out low-expressed genes).
We further lter out 8,000 genes with smaller variance (i.e. genes with little predictive information).
Finally, the remaining 2,000 genes are used in the analysis.
}
\references{
Kohlmann, A., Kipps, T. J., Rassenti, L. Z., Downing, J. R., Shurtleff, S. A., Mills, K. I., Gilkes,
A. F., Hofmann, W.-K., Basso, G., DellOrto, M. C., et al. (2008). An international standardization
programme towards the application of gene expression profiling in routine leukaemia diagnostics:
 the microarray innovations in leukemia study prephase. British journal of haematology,142(5):802-807.
}
\keyword{datasets}
