% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds.GSE17855}
\alias{ds.GSE17855}
\title{Microarray data of leukemia after preprocessing(GSE17855)}
\format{
a list of two components. Expression matrix with 74 samples(Row) and 2000 genes(columns); A vector of class label for samples.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE17855}
}
\usage{
ds.GSE17855
}
\description{
Expression matrix with 74 samples(Row) and 2000 genes(columns). Originally there are 54,676 probesets in each dataset
(GSE6891, GSE17855 and GSE13159) and we remove the probesets with missing values and select the probesets with the largest interquartile
range to represent the gene if multiple probesets are mapped to the same gene. 20,192 unique genes
remained for every study after this preprocessing. Furthermore, for each study, we transform data
to log scale and only keep the top 10,000 genes with the largest mean expression level (i.e. filter out low-expressed genes).
We further lter out 8,000 genes with smaller variance (i.e. genes with little predictive information).
Finally, the remaining 2,000 genes are used in the analysis.
}
\references{
Balgobind, B. V., Van den Heuvel-Eibrink, M. M., Menezes, R. X., Reinhardt, D., Hollink, I. H.,
Peters, S. T., van Wering, E. R., Kaspers, G. J., Cloos, J., de Bont, E. S. M., et al. (2010).
Evaluation of gene expression signatures predictive for cytogenetic and molecular subtypes of
pediatric acute myeloid leukemia. haematologica, pages haematol-2010.
}
\keyword{datasets}
