% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds.GSE47474}
\alias{ds.GSE47474}
\title{RNA Sequencing data of rat brain after preprocessing(GSE47474)}
\format{
A list of two components: Expression matrix with 36 samples(Row) and 2000 genes(columns); a vector of class label for samples.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE47474}
}
\usage{
ds.GSE47474
}
\description{
Expression matrix with 36 samples(Row) and 2000 genes(columns). RNA samples from three brain regions (hippocampus, striatum and
prefrontal cortex) were sequenced for both control strains and HIV strains. Only the 36 control
strains (12 in each brain region) are used here to see whether samples from three brain regions can
be correctly clustered The original count data is transformed into
CPM value and then 2000 genesare kept by filtering low-expressed genes and low-variance genes.
}
\references{
Li, M. D., Cao, J., Wang, S., Wang, J., Sarkar, S., Vigorito, M., Ma, J. Z., and Chang, S. L. (2013).
Transcriptome sequencing of gene expression in the brain of the hiv-1 transgenic rat. PLoS One, 8(3):e59582.
}
\keyword{datasets}
