% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds.ISOLET}
\alias{ds.ISOLET}
\title{letter-name dataset after preprocessing}
\format{
A list of two components: Matrix with 617 samples(Row) and 1200 features(columns); a vector of class label for samples.
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/isolet}
}
\usage{
ds.ISOLET
}
\description{
Matrix with 617 samples(Row) and 1200 features(columns). ISOLET dataset was generated by a study where 150 subjects spoke each letter
of the alphabet twice and recorded 617 features including spectral coefficients, contour features,
sonorant features, pre-sonorant features and post-sonorant features. We only use five vowels and
1200 training subjects ( 240 samples for each of five vowels).
}
\references{
Dheeru, D. and Karra Taniskidou, E. (2017). UCI machine learning repository.
}
\keyword{datasets}
