% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds.Leaf}
\alias{ds.Leaf}
\title{Plant species leaves dataset after preprocessing}
\format{
A list of two components: Matrix with 64 samples(Row) and 190 features(columns); a vector of class label for samples.
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/One-hundred+plant+species+leaves+data+set}
}
\usage{
ds.Leaf
}
\description{
Matrix with 64 samples(Row) and 190 features(columns). Mallah et al. (2013) introduced a dataset consisting of one-hundred
species of plants with three types of features for leaves: shape texture and margin. Here we
only consider 4 species out of 100, alnus rubra, acer capillipes, cornus macrophylla and quercus
chrysolepis. After deleting features with any missing values, we have 64 samples (16 for each species) and 190 features.
}
\references{
Mallah, C., Cope, J., and Orwell, J. (2013). Plant leaf classification using probabilistic integration
of shape, texture and margin features. Signal Processing, Pattern Recognition and Applications, 5(1).
}
\keyword{datasets}
