% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds.Pancancer}
\alias{ds.Pancancer}
\title{RNA Sequencing data after preprocessing(Pancancer)}
\format{
A list of two components: Expression matrix with 801 samples(Row) and 2000 genes(columns).
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/gene+expression+cancer+RNA-Seq}
}
\usage{
ds.Pancancer
}
\description{
Expression matrix with 801 samples(Row) and 2000 genes(columns). This collection of data consists of five different
types of tumor: 300 breast cancer (BRCA), 146 kidney clear cell carcinoma (KIRC), 78 colon
cancer (COAD), 141 lung adenocarcinoma (LUAD) and 136 prostate cancer (PRAD). The data
has already been normalized and we use the same filtering process to keep 2000 genes; a vector of class label for samples.
}
\keyword{datasets}
