% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds.TissueType}
\alias{ds.TissueType}
\title{Microarray data after preprocessing(Mammalian tissue types dataset:)}
\format{
A list of two components: Expression matrix with 102 samples(Row) and 2000 genes(columns); a vector of class label for samples.
}
\source{
\url{http://portals.broadinstitute.org/cgi-bin/cancer/datasets.cgi}
}
\usage{
ds.TissueType
}
\description{
Expression matrix with 102 samples(Row) and 2000 genes(columns). Gene expression from human and mouse samples across a diverse
array of tissues, organs, and cell lines have been profiled by Su et al. (2002). Here we
only consider four tissue types: breast, prostate, lung, and colon.
The original data has 102 samples and 5565 probesets (genes). Following
similar proprocessing procedure, We filter 3000 genes with highest expression value and then 2000
genes were used in the analysis after further filtering low-variance genes.
}
\references{
Su, A. I., Cooke, M. P., Ching, K. A., Hakak, Y., Walker, J. R., Wiltshire, T., Orth, A. P., Vega,
R. G., Sapinoso, L. M., Moqrich, A., et al. (2002). Large-scale analysis of the human and mouse
transcriptomes. Proceedings of the National Academy of Sciences, 99(7):4465-4470.
}
\keyword{datasets}
