% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K.Clust.R
\name{K.Clust}
\alias{K.Clust}
\title{Estimate number of clusters for K-Means}
\usage{
K.Clust(
  data,
  method = "S4",
  Kmin = 2,
  Kmax = 10,
  trim.S4 = 0.05,
  cutoff = 0.8,
  n.resample = 50
)
}
\arguments{
\item{data}{A matrix where rows represent samples and columns represent features}

\item{method}{Method to use, including all the methods compared in S4 paper. Please refer to the reference of S4 paper
\itemize{
\item{"S4": }{Simultaneous Estimation of Number of Clusters and Feature Sparsity parameter in High-Dimensional Data Clustering Analysis. Biometrics accepted}
\item{"Gap.PCA", "Gap.Unif": }{Tibshirani, R., Walther, G., and Hastie, T. (2001).}
\item{"PS": }{Tibshirani, R. and Walther, G. (2005). }
\item{"Jump": }{Sugar, C. A. and James, G. M. (2003)}
\item{"CH": }{Calinski, T. and Harabasz, J. (1974)}
\item{"FW": }{Fang, Y. and Wang, J. (2012)}
\item{"LD": }{Levine, E. and Domany, E. (2001).}
\item{"KL": }{Krzanowski, W. J. and Lai, Y. (1988).}
\item{"H": }{Hartigan, J. A. (1975). }
\item{"silhouette": }{Rousseeuw, P. J. (1987).}
}}

\item{Kmin}{integer. Minimum number of clusters}

\item{Kmax}{integer. Maximum number of clusters.}

\item{trim.S4}{The precentage of trim for S4 method}

\item{cutoff}{The threshold for detecting null cluster structure for resampling-based methods. If the maximum score for all K is smaller than cutoff, return one cluster.}

\item{n.resample}{Number of resamples in resampling-based methods. Also for Gap statistic, n.resample means the number of reference data to generate.}
}
\value{
For S4 method, "maxk" is the optimal number of clusters and "score" is the stability score for K from kmin to kmax. For Gap, PS, Jump and FW, "maxk" is the estimated number of clusters and model is the original output (These methods are implemented internally using other packages (package "fpc" and package "cluster", readers can also read the reference for details). For other methods, only optimal number of clusters is returned.
}
\description{
Estimate number of clusters for K-Means
}
\examples{
\dontrun{
data<-Sim1(settings = "A1")
#using S4
res.S4<-K.Clust(data$x,method="S4",Kmin=2,Kmax=10,trim.S4=0.05,cutoff=0.8,n.resample=50)
#Other methods Take prediction strength for an example.
res.PS<-K.Clust(data$x,method="PS",Kmin=2,Kmax=10,trim.S4=0.05,cutoff=0.8,n.resample=50)
}
}
\references{
S4: Simultaneous Estimation of Number of Clusters and Feature Sparsity parameter in High-Dimensional Data Clustering Analysis. Biometrics accepted.

"Gap.PCA" and "Gap.Unif": Tibshirani, R., Walther, G., and Hastie, T. (2001). Estimating the number of clusters in a data set via the gap statistic. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 63(2):411-423.

"PS":  Tibshirani, R. and Walther, G. (2005). Cluster validation by prediction strength. Journal of Computational and Graphical Statistics, 14(3):511-528.

"Jump":  Sugar, C. A. and James, G. M. (2003). Finding the number of clusters in a dataset: An informationtheoretic approach. Journal of the American Statistical Association, 98(463):750-763.

"CH":  Calinski, T. and Harabasz, J. (1974). A dendrite method for cluster analysis. Communications in Statistics-theory and Methods, 3(1):1-27.

"FW":  Fang, Y. and Wang, J. (2012). Selection of the number of clusters via the bootstrap method. Computational Statistics & Data Analysis, 56(3):468-477.

"LD":  Levine, E. and Domany, E. (2001). Resampling method for unsupervised estimation of cluster validity. Neural computation, 13(11):2573{2593.

"KL":  Krzanowski, W. J. and Lai, Y. (1988). A criterion for determining the number of groups in a dataset using sum-of-squares clustering. Biometrics, pages 23-34.

"H":  Hartigan, J. A. (1975). Clustering algorithms, new york: John willey and sons. Inc. Pages113129.}

"silhouette":  Rousseeuw, P. J. (1987). Silhouettes: a graphical aid to the interpretation and validation of cluster analysis. Journal of computational and applied mathematics, 20:53-65.
}
