% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL.PS.R
\name{KL.PS}
\alias{KL.PS}
\title{Estimate number of clusters and sparsity parameter simultaneously by extended prediction strength.}
\usage{
KL.PS(
  x,
  lambda_list,
  cv = 2,
  k_vector = c(2, 3, 4),
  M = 20,
  cutoff,
  num.cores = 1
)
}
\arguments{
\item{x}{data matrix, where rows represent samples and columns represents features}

\item{lambda_list}{a list whose length is equal to k_vector. Every element is a vector of lambda
as the search region for this K.}

\item{cv}{Number of cross validation to do in each resampling. cv=2 is suggested by
Tibshirani, R. and Walther, G. (2005)}

\item{k_vector}{The search pool for number of clusters}

\item{M}{Number of resampling. Every resampling is to divide the data into train and half.}

\item{num.cores}{Number of cpus used in parallel computing. It is suggest to run large num_cores
in servers instead of laptops.}
}
\value{
A list of two components:
\itemize{
\item{res_k: }{Optimal number of clusters}
\item{res_w: }{Optimal lambda}

}
}
\description{
Estimate number of clusters and sparsity parameter simultaneously by extended prediction strength.
}
\examples{
\dontrun{
#generate simulation II data
data<-Sim2(h=200,q = 50,u=0.8)#For demo purpose, we use 200 features in total for fast result.
#using Efficient Algorithm for Choosing Grids of lambda for each K
k_vector<-2:7#search K from 2 to 7
wbounds_list<-list(1)
for(l in 1:length(k_vector)){#for each K, using the algorithm to get 20 lambda.
 wbounds_list[[l]] = region.lambda(lam1=1.5,iteration=20,data,k_vector[l])
}
#This part of code is to delete the lambda which selecting all the genes.
#since our S4 method calculate specificity, the lambda selecting all the genes must be removed.
for(l in 1:length(k_vector)){
 temp<-KMeansSparseCluster(data,K=k_vector[l],wbounds=wbounds_list[[l]],nstart=100)
 num<-rep(0,length(temp))
 for(i in 1:length(num)){
   num[i]<-sum(temp[[i]]$ws>0)
 }
 if(sum(num==ncol(data))>0){
   wbounds_list[[l]]<-wbounds_list[[l]][1:(min(which(num==ncol(data)))-3)]
 }
}
#run extended Prediction strength method
res.PS<-KL.PS(x=data,lambda_list = wbounds_list,k_vector = k_vector,cv = 2,M=20,num.cores = 1,cutoff=0.8)
}
}
\references{
Tibshirani, R. and Walther, G. (2005). Cluster validation by prediction strength. Journal of Computational and Graphical Statistics, 14(3):511-528.
}
