\docType{class}
\name{profileR-class}
\alias{profileR-class}
\title{Class "profileR" for the methods in profileR}
\description{
  The output created by \code{\link{profile.reliability}},
  \code{\link{profile.by.group}}, and \code{\link{criterion.pattern}} 
  is represented as a \code{\linkS4class{profileR}} object.
}
\section{Objects from the "profileR" Class}{
  Objects are created by calls to
  \code{\link{profile.reliability}},
  \code{\link{profile.by.group}} or
  \code{\link{criterion.pattern}}.
}
\examples{
showClass("profileR")
methods(class="profileR")
}
\seealso{
  \code{\link{profile.reliability}},
  \code{\link{profile.by.group}},
  \code{\link{criterion.pattern}},
}
\keyword{classes}

