% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.ranks.R
\name{random.ranks}
\alias{random.ranks}
\title{Function to assign the random region for a expression matrix}
\usage{
random.ranks(bs, n = 1000, seed = 123456, mc.cores = 2, renice = T)
}
\arguments{
\item{bs}{Full return data returned from the sig.ranksum function.}

\item{n}{Number of random samples.}

\item{seed}{Seed to generate random data (for reproducibility).}

\item{mc.cores}{An integer specifying how many cores to use.}

\item{renice}{A boolean to specify if the child processes should be reniced to the lowest nice priority.}
}
\description{
This is a function to assign the random region for a dataset
}

