% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.lambda.R
\name{region.lambda}
\alias{region.lambda}
\title{Efficient Algorithm for Choosing Grids of lambda}
\usage{
region.lambda(lam1 = 3, iteration = 30, x, K)
}
\arguments{
\item{lam1}{The lower bound for lambda}

\item{iteration}{Number of iterative search}

\item{x}{data matrix, where rows represent samples and columns represents features}

\item{K}{Number of clusters}
}
\value{
A vector of sorted lambda value
}
\description{
Efficient Algorithm for Choosing Grids of lambda
}
\examples{
\dontrun{
data(data_tanbin_mat)
data<-RatBrain$data
region.lambda(lam1=3,iteration=20,x=data,K=3)
}
}
