% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim1.R
\name{Sim1}
\alias{Sim1}
\title{Function to simulate data for 12 settings in simulation I. A1-A7 are well-separated settings and B1-B5 are not-well-separated settings, in terms of ARI and difficulty to separate them from Null data.}
\usage{
Sim1(settings = "A1")
}
\arguments{
\item{settings}{Simulate setting from 1 to 10 corresponding to the simulation1 in S4 paper.}
}
\value{
The data matrix corresponding to each setting
}
\description{
Function to simulate data for 12 settings in simulation I.
}
\examples{
\dontrun{
Sim1(settings="2")
}
}
