% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim2.R
\name{Sim2}
\alias{Sim2}
\title{Simulate high-dimensional data with independent feature structure of three clusters.}
\usage{
Sim2(h, q, u)
}
\arguments{
\item{h}{Total number of features}

\item{q}{Number of informative features}

\item{u}{Effect size}
}
\value{
A matrix with 99 rows(samples) and h column(features),
Sample 1-33 is cluster 1, 34-66 is cluster 2 and 67-99 is cluster 3.
The first q features are informative fearures whereas others are random noise
}
\description{
Simulate high-dimensional data with independent feature structure of three clusters. (Simulation II)
}
\examples{
\dontrun{
h=1000
q=200
u=0.8
Sim2(h=h,q=200,u=0.8)
}
}
\references{
Manuscript: Simultaneous Estimation of Number of Clusters and Feature Sparsity in Clustering High-Dimensional Data Using Resampling
}
