% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim3.R
\name{Sim3}
\alias{Sim3}
\title{Simulate high-dimensional data with correlated feature structure of three clusters (Simulation III).}
\usage{
Sim3(
  POIN1 = POIN1,
  POIN2 = POIN2,
  POIN3 = POIN3,
  Nm = POIM,
  M,
  Ua,
  Ub,
  Ulow,
  Uupper,
  sigma1,
  sigma2,
  Nnoise = 600,
  cov = 0.5
)
}
\arguments{
\item{POIN1}{Number of samples in cluster 1 follows possion distribution. POIN1 is possion mean.}

\item{POIN2}{Number of samples in cluster 2 follows possion distribution. POIN2 is possion mean.}

\item{POIN3}{Number of samples in cluster 3 follows possion distribution. POIN3 is possion mean.}

\item{Nm}{Number of genes within each gene module follows possion distribution. Nm is possion mean.}

\item{M}{Number of gene modules of informative genes.}

\item{Ua, Ub}{Template gene expression of each cluster for each informative gene is simulated from Unif(Ua,Ub)}

\item{Ulow, Uupper}{Effect size(The constraint for template gene expression).
The template gene expression of each two clusters for every gene will be between Ulow and Uupper, where Ulow is lower bound.}

\item{sigma1}{Biological variation to template gene expression of informative genes}

\item{sigma2}{Biological variation to template gene expression of noise genes}

\item{Nnoise}{Number of noise genes.}

\item{cov}{Determine the correlation of genes within each module. The higher cov, the larger correlation.}
}
\value{
A list of three components:
\itemize{
\item{data: }{A data matrix where row represents samples and column represents features}
\item{true_label: }{class label}
\item{true_feature: }{a vector equal to number of columns of data, where 1 represents informative features and 0 otherwise.}
}
}
\description{
Simulate high-dimensional data with correlated feature structure of three clusters (Simulation III).
}
\examples{
\dontrun{
#####this is the default hyperparameter in the simulation3 of manuscript.
POIN1<-40
POIN2<-30
POIN3<-20
sigma1<-0.2
sigma2<-1
POIM<-20
Ua<-4
Ub<-10
j1<-1
j2<-1
Nnoise<-600

#################below are the tuning parameter
cov=0.1
Ulow=1.0
Uupper=1.5
M=10
Sim3(POIN1=POIN1,POIN2=POIN2,POIN3=POIN3,Nm=POIM,
M=M,Ua=Ua,Ub=Ub,Ulow=Ulow,Uupper=Uupper,sigma1=sigma1,sigma2=sigma2,Nnoise=Nnoise,cov=cov)
}
}
\references{
Manuscript: Simultaneous Estimation of Number of Clusters and Feature Sparsity in Clustering High-Dimensional Data Using Resampling.
}
